application/src/main/java/com/hanfak/airport/infrastructure/webserver/notfound/NotFoundServlet.java
package com.hanfak.airport.infrastructure.webserver.notfound;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

public class NotFoundServlet extends HttpServlet {
    // TODO return available endpoints
    private void writeNotFoundPage(HttpServletResponse response) throws IOException {
        response.setContentType("text/html");
        response.setStatus(404);
        response.getWriter().write("Not available");
    }

    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }

    @Override
    protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }

    @Override
    protected void doHead(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }

    @Override
    protected void doOptions(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }

    @Override
    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }

    @Override
    protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }

    @Override
    protected void doTrace(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        writeNotFoundPage(response);
    }
}
